Separated some utility functions and configurations into different files
