+++ main/AutoSubSync.py
@@ -5,48 +5,32 @@ import shutil
 import subprocess
 import threading
 import xml.etree.ElementTree as ET
-import ctypes
 import json
 import webbrowser
-import platform
+# Tkinter GUI
 import tkinter as tk
 import tkinter.font as tkFont
 from datetime import datetime
 from tkinter import filedialog, ttk, messagebox, PhotoImage, Menu, simpledialog
 from tkinterdnd2 import DND_FILES, TkinterDnD
-from alass_encodings import enc_list
-import cchardet
-import charset_normalizer
-import chardet
 import requests
 import psutil
 import signal
-import texts
+from alass_encodings import enc_list
+from functions.get_platform import platform
+from constants import *
+from theme import *
+from texts_constants import *
+from utils import *
+from config import VERSION, config
+from functions.get_config import config_dir, config_path
+from functions.get_desktop_path import desktop_path
 
-platform = platform.system()
 if platform == "Darwin":  # macOS
     from tkmacosx import Button as TkButton
 else:  # Windows or Linux
     from tkinter import Button as TkButton
 
-
-def get_base_dir():
-    if getattr(sys, "frozen", False):
-        # Running as PyInstaller executable
-        if platform == "Linux":
-            base_dir = os.path.expanduser("~/.AutoSubSync")
-            # Ensure the directory exists
-            if not os.path.exists(base_dir):
-                os.makedirs(base_dir, exist_ok=True)
-        else:
-            base_dir = os.path.dirname(sys.executable)
-    else:
-        # Running as Python script
-        base_dir = os.path.dirname(os.path.abspath(__file__))
-    return base_dir
-
-
-base_dir = get_base_dir()
 program_dir = os.path.dirname(os.path.abspath(__file__))
 
 # Set the working directory to the script's directory
@@ -112,100 +96,6 @@ if platform in ["Darwin", "Linux"]:
     if errors:
         messagebox.showerror("Permission Error", "\n".join(errors))
 
-# Config file
-default_settings = {
-    "theme": "system",
-    "keep_logs": True,
-    "log_window_font": "Cascadia Code SemiLight",
-    "log_window_font_size": 7,
-    "log_window_font_style": "normal",
-    "remember_the_changes": True,
-    "notify_about_updates": True,
-    "ffsubsync_option_framerate": False,
-    "ffsubsync_option_gss": False,
-    "ffsubsync_option_vad": "default",
-    "alass_disable_fps_guessing": False,
-    "alass_speed_optimization": False,
-    "alass_split_penalty": 7,
-    "action_var_auto": "OPTION_SAVE_NEXT_TO_SUBTITLE",
-    "sync_tool_var_auto": "SYNC_TOOL_FFSUBSYNC",
-    "keep_converted_subtitles": True,
-    "keep_extracted_subtitles": True,
-    "backup_subtitles_before_overwriting": True,
-    "check_video_for_subtitle_stream_in_alass": True,
-    "add_prefix": True,
-    "additional_ffsubsync_args": "",
-    "additional_alass_args": "",
-}
-
-config_path = os.path.join(base_dir, "config.json")
-
-
-def create_config_file():
-    with open(config_path, "w", encoding="utf-8") as config_file:
-        json.dump(default_settings, config_file, indent=4)
-
-
-if not os.path.exists(config_path):
-    create_config_file()
-
-try:
-    with open(config_path, "r", encoding="utf-8") as config_file:
-        config = json.load(config_file)
-except FileNotFoundError:
-    config = {}
-    messagebox.showerror("Error", '"config.json" file not found')
-try:
-    with open("VERSION", "r", encoding="utf-8") as version_file:
-        VERSION = version_file.read().strip()
-except FileNotFoundError:
-    VERSION = " UNKNOWN VERSION"
-    messagebox.showerror("Error", '"VERSION" file not found')
-
-
-def is_dark_mode():
-    if platform == "Windows":
-        try:
-            import winreg
-
-            with winreg.OpenKey(
-                winreg.HKEY_CURRENT_USER,
-                r"Software\Microsoft\Windows\CurrentVersion\Themes\Personalize",
-            ) as key:
-                value, _ = winreg.QueryValueEx(key, "AppsUseLightTheme")
-                return value == 0
-        except Exception:
-            return False
-    elif platform == "Linux":
-        # Check for GNOME dark mode
-        try:
-            result = subprocess.run(
-                ["gsettings", "get", "org.gnome.desktop.interface", "gtk-theme"],
-                capture_output=True,
-                text=True,
-                check=False,
-            )
-            return "dark" in result.stdout.lower()
-        except Exception:
-            return False
-    return False
-
-
-# Font Fix: if platform is macOS or Linux, adjust font settings accordingly
-if platform == "Darwin" and config.get("log_window_font", "Cascadia Code") in [
-    "Cascadia Code",
-    "Cascadia Code SemiLight",
-]:
-    config["log_window_font"] = "Andale Mono"
-    config["log_window_font_size"] = 10
-    config["log_window_font_style"] = "normal"
-elif platform == "Linux" and config.get("log_window_font", "Cascadia Code") in [
-    "Cascadia Code",
-    "Cascadia Code SemiLight",
-]:
-    config["log_window_font"] = "Noto Sans Sinhala"
-    config["log_window_font_size"] = 8
-    config["log_window_font_style"] = "normal"
 # Save the updated configuration
 try:
     with open(config_path, "w", encoding="utf-8") as config_file:
@@ -213,475 +103,9 @@ try:
 except Exception as e:
     messagebox.showerror("Error", "Failed to fix font: " + str(e))
 
-# Fix small font size on macOS
-if platform == "Darwin":  # macOS
-    FONT_SIZE = 12  # Bigger font for macOS
-    FONT_SIZE_TWO = 14
-else:  # Windows or Linux
-    FONT_SIZE = 8  # Default font size
-    FONT_SIZE_TWO = 10
-
-# Get the theme from the config, or set it based on the system's theme
-THEME = config.get("theme", "system")
-if THEME == "system":
-    THEME = "dark" if is_dark_mode() else "light"
-
-# File extensions
-FFSUBSYNC_SUPPORTED_EXTENSIONS = [".srt", ".ass", ".ssa", ".vtt"]
-ALASS_SUPPORTED_EXTENSIONS = [".srt", ".ass", ".ssa", ".sub", ".idx"]
-ALASS_EXTRACTABLE_SUBTITLE_EXTENSIONS = {"subrip": "srt", "ass": "ass", "webvtt": "vtt"}
-SUBTITLE_EXTENSIONS = [
-    ".srt",
-    ".vtt",
-    ".sbv",
-    ".sub",
-    ".ass",
-    ".ssa",
-    ".dfxp",
-    ".ttml",
-    ".itt",
-    ".stl",
-    ".idx",
-]
-VIDEO_EXTENSIONS = [
-    ".mp4",
-    ".mkv",
-    ".avi",
-    ".webm",
-    ".flv",
-    ".mov",
-    ".wmv",
-    ".mpg",
-    ".mpeg",
-    ".m4v",
-    ".3gp",
-    ".h264",
-    ".h265",
-    ".hevc",
-]
-# Define color schemes for light and dark themes
-COLOR_SCHEMES = {
-    "COLOR_BW": ("black", "white"),
-    "COLOR_WB": ("white", "black"),
-    "COLOR_BACKGROUND": ("#f1f0f1", "#202020"),
-    "COLOR_PRIMARY": ("gray85", "gray10"),
-    "COLOR_SECONDARY": ("gray75", "gray30"),
-    "COLOR_TAB_INACTVE": ("gray20", "gray60"),
-    "COLOR_PROGRESSBAR": ("#00a31e", "#00a31e"),
-    "COLOR_ZERO": ("lightgrey", "grey10"),
-    "COLOR_ONE": ("lightgrey", "grey20"),
-    "COLOR_TWO": ("lightgreen", "#0a420a"),
-    "COLOR_THREE": ("lightblue", "#12303b"),
-    "COLOR_FOUR": ("lightgoldenrodyellow", "#5c5c0a"),
-    "COLOR_FIVE": ("lightcoral", "#5b0b0b"),
-    "COLOR_SIX": ("RosyBrown1", "red4"),
-    "COLOR_SEVEN": ("#007FFF", "#389afc"),
-    "COLOR_EIGHT": ("red", "#ffeded"),
-    "COLOR_NINE": ("green", "lightgreen"),
-    "COLOR_OPTIONS": ("gray85", "gray30"),
-    "TREEVIEW_SELECTED_COLOR": ("steelblue", "#325d81"),
-    "COLOR_MILISECONDS_HIGH": ("aliceblue", "#001b33"),
-    "COLOR_MILISECONDS_LOW": ("mistyrose", "#330500"),
-    "DEFAULT_BUTTON_COLOR": ("gray50", "gray50"),
-    "DEFAULT_BUTTON_COLOR_ACTIVE": ("gray40", "gray40"),
-    "BUTTON_COLOR_MANUAL": ("#32993a", "#3ec149"),
-    "BUTTON_COLOR_MANUAL_ACTIVE": ("#2d8a35", "#38ad42"),
-    "BUTTON_COLOR_AUTO": ("royalblue", "#6699ff"),
-    "BUTTON_COLOR_AUTO_ACTIVE": ("RoyalBlue3", "#6585e7"),
-    "BUTTON_COLOR_BATCH": ("#b05958", "#be7674"),
-    "BUTTON_COLOR_BATCH_ACTIVE": ("#a15150", "#b66463"),
-}
-# Select the appropriate color scheme based on the theme
-is_dark_theme = THEME == "dark"
-border_fix = 0 if is_dark_theme else 1
-COLOR_BACKGROUND = COLOR_SCHEMES["COLOR_BACKGROUND"][is_dark_theme]
-COLOR_BW = COLOR_SCHEMES["COLOR_BW"][is_dark_theme]
-COLOR_WB = COLOR_SCHEMES["COLOR_WB"][is_dark_theme]
-COLOR_PRIMARY = COLOR_SCHEMES["COLOR_PRIMARY"][is_dark_theme]
-COLOR_SECONDARY = COLOR_SCHEMES["COLOR_SECONDARY"][is_dark_theme]
-COLOR_TAB_INACTVE = COLOR_SCHEMES["COLOR_TAB_INACTVE"][is_dark_theme]
-COLOR_PROGRESSBAR = COLOR_SCHEMES["COLOR_PROGRESSBAR"][is_dark_theme]
-COLOR_ZERO = COLOR_SCHEMES["COLOR_ZERO"][is_dark_theme]
-COLOR_ONE = COLOR_SCHEMES["COLOR_ONE"][is_dark_theme]
-COLOR_TWO = COLOR_SCHEMES["COLOR_TWO"][is_dark_theme]
-COLOR_THREE = COLOR_SCHEMES["COLOR_THREE"][is_dark_theme]
-COLOR_FOUR = COLOR_SCHEMES["COLOR_FOUR"][is_dark_theme]
-COLOR_FIVE = COLOR_SCHEMES["COLOR_FIVE"][is_dark_theme]
-COLOR_SIX = COLOR_SCHEMES["COLOR_SIX"][is_dark_theme]
-COLOR_SEVEN = COLOR_SCHEMES["COLOR_SEVEN"][is_dark_theme]
-COLOR_EIGHT = COLOR_SCHEMES["COLOR_EIGHT"][is_dark_theme]
-COLOR_NINE = COLOR_SCHEMES["COLOR_NINE"][is_dark_theme]
-COLOR_OPTIONS = COLOR_SCHEMES["COLOR_OPTIONS"][is_dark_theme]
-TREEVIEW_SELECTED_COLOR = COLOR_SCHEMES["TREEVIEW_SELECTED_COLOR"][is_dark_theme]
-COLOR_MILISECONDS_HIGH = COLOR_SCHEMES["COLOR_MILISECONDS_HIGH"][is_dark_theme]
-COLOR_MILISECONDS_LOW = COLOR_SCHEMES["COLOR_MILISECONDS_LOW"][is_dark_theme]
-DEFAULT_BUTTON_COLOR = COLOR_SCHEMES["DEFAULT_BUTTON_COLOR"][is_dark_theme]
-DEFAULT_BUTTON_COLOR_ACTIVE = COLOR_SCHEMES["DEFAULT_BUTTON_COLOR_ACTIVE"][
-    is_dark_theme
-]
-BUTTON_COLOR_MANUAL = COLOR_SCHEMES["BUTTON_COLOR_MANUAL"][is_dark_theme]
-BUTTON_COLOR_MANUAL_ACTIVE = COLOR_SCHEMES["BUTTON_COLOR_MANUAL_ACTIVE"][is_dark_theme]
-BUTTON_COLOR_AUTO = COLOR_SCHEMES["BUTTON_COLOR_AUTO"][is_dark_theme]
-BUTTON_COLOR_AUTO_ACTIVE = COLOR_SCHEMES["BUTTON_COLOR_AUTO_ACTIVE"][is_dark_theme]
-BUTTON_COLOR_BATCH = COLOR_SCHEMES["BUTTON_COLOR_BATCH"][is_dark_theme]
-BUTTON_COLOR_BATCH_ACTIVE = COLOR_SCHEMES["BUTTON_COLOR_BATCH_ACTIVE"][is_dark_theme]
-# Language selection (ALL TRANSLATIONS ARE LOCATED IN "texts.py")
-LANGUAGE = config.get("language", "en")
-LANGUAGES = {
-    "English": "en",
-    "Español": "es",
-    "Deutsch": "de",
-    "Français": "fr",
-    "Italiano": "it",
-    "Polski": "pl",
-    "Português": "pt",
-    "Türkçe": "tr",
-    "Tiếng Việt": "vi",
-    "Bahasa Indonesia": "id",
-    "Bahasa Melayu": "ms",
-    "ไทย": "th",
-    "Українська": "uk",
-    "Русский": "ru",
-    "中国人": "zh",
-    "日本語": "ja",
-    "한국어": "ko",
-    "हिन्दी": "hi",
-    "বাংলা": "bn",
-    "اردو": "ur",
-    "العربية": "ar",
-    "فارسی": "fa",
-}
-# Program information
-PROGRAM_NAME = texts.PROGRAM_NAME
-GITHUB_URL = texts.GITHUB_URL
-GITHUB_VERSION_URL = texts.GITHUB_VERSION_URL
-GITHUB_LATEST_RELEASE_URL = texts.GITHUB_LATEST_RELEASE_URL
-# Tooltip texts for checkboxes
-TOOLTIP_SAVE_TO_DESKTOP = texts.TOOLTIP_SAVE_TO_DESKTOP[LANGUAGE]
-TOOLTIP_REPLACE_ORIGINAL = texts.TOOLTIP_REPLACE_ORIGINAL[LANGUAGE]
-TOOLTIP_GSS = texts.TOOLTIP_GSS[LANGUAGE]
-TOOLTIP_VAD = texts.TOOLTIP_VAD[LANGUAGE]
-TOOLTIP_FRAMERATE = texts.TOOLTIP_FRAMERATE[LANGUAGE]
-TOOLTIP_ALASS_SPEED_OPTIMIZATION = texts.TOOLTIP_ALASS_SPEED_OPTIMIZATION[LANGUAGE]
-TOOLTIP_ALASS_DISABLE_FPS_GUESSING = texts.TOOLTIP_ALASS_DISABLE_FPS_GUESSING[LANGUAGE]
-TOOLTIP_TEXT_ACTION_MENU_AUTO = texts.TOOLTIP_TEXT_ACTION_MENU_AUTO[LANGUAGE]
-TOOLTIP_TEXT_SYNC_TOOL_MENU_AUTO = texts.TOOLTIP_TEXT_SYNC_TOOL_MENU_AUTO[LANGUAGE]
-# Dialogs - Buttons
-NOTIFY_ABOUT_UPDATES_TEXT = texts.NOTIFY_ABOUT_UPDATES_TEXT[LANGUAGE]
-UPDATE_AVAILABLE_TEXT = texts.UPDATE_AVAILABLE_TEXT[LANGUAGE]
-UPDATE_AVAILABLE_TITLE = texts.UPDATE_AVAILABLE_TITLE[LANGUAGE]
-LANGUAGE_LABEL_TEXT = texts.LANGUAGE_LABEL_TEXT[LANGUAGE]
-TAB_AUTOMATIC_SYNC = texts.TAB_AUTOMATIC_SYNC[LANGUAGE]
-TAB_MANUAL_SYNC = texts.TAB_MANUAL_SYNC[LANGUAGE]
-CANCEL_TEXT = texts.CANCEL_TEXT[LANGUAGE]
-GENERATE_AGAIN_TEXT = texts.GENERATE_AGAIN_TEXT[LANGUAGE]
-GO_BACK = texts.GO_BACK[LANGUAGE]
-BATCH_MODE_TEXT = texts.BATCH_MODE_TEXT[LANGUAGE]
-NORMAL_MODE_TEXT = texts.NORMAL_MODE_TEXT[LANGUAGE]
-START_AUTOMATIC_SYNC_TEXT = texts.START_AUTOMATIC_SYNC_TEXT[LANGUAGE]
-START_BATCH_SYNC_TEXT = texts.START_BATCH_SYNC_TEXT[LANGUAGE]
-BATCH_INPUT_TEXT = texts.BATCH_INPUT_TEXT[LANGUAGE]
-BATCH_INPUT_LABEL_TEXT = texts.BATCH_INPUT_LABEL_TEXT[LANGUAGE]
-SHIFT_SUBTITLE_TEXT = texts.SHIFT_SUBTITLE_TEXT[LANGUAGE]
-LABEL_SHIFT_SUBTITLE = texts.LABEL_SHIFT_SUBTITLE[LANGUAGE]
-REPLACE_ORIGINAL_TITLE = texts.REPLACE_ORIGINAL_TITLE[LANGUAGE]
-REPLACE_ORIGINAL_TEXT = texts.REPLACE_ORIGINAL_TEXT[LANGUAGE]
-FILE_EXISTS_TITLE = texts.FILE_EXISTS_TITLE[LANGUAGE]
-FILE_EXISTS_TEXT = texts.FILE_EXISTS_TEXT[LANGUAGE]
-ALREADY_SYNCED_FILES_TITLE = texts.ALREADY_SYNCED_FILES_TITLE[LANGUAGE]
-ALREADY_SYNCED_FILES_MESSAGE = texts.ALREADY_SYNCED_FILES_MESSAGE[LANGUAGE]
-SUBTITLE_INPUT_TEXT = texts.SUBTITLE_INPUT_TEXT[LANGUAGE]
-VIDEO_INPUT_TEXT = texts.VIDEO_INPUT_TEXT[LANGUAGE]
-LABEL_DROP_BOX = texts.LABEL_DROP_BOX[LANGUAGE]
-WARNING = texts.WARNING[LANGUAGE]
-CONFIRM_RESET_MESSAGE = texts.CONFIRM_RESET_MESSAGE[LANGUAGE]
-TOGGLE_KEEP_CONVERTED_SUBTITLES_WARNING = texts.TOGGLE_KEEP_CONVERTED_SUBTITLES_WARNING[
-    LANGUAGE
-]
-TOGGLE_KEEP_EXTRACTED_SUBTITLES_WARNING = texts.TOGGLE_KEEP_EXTRACTED_SUBTITLES_WARNING[
-    LANGUAGE
-]
-BACKUP_SUBTITLES_BEFORE_OVERWRITING_WARNING = (
-    texts.BACKUP_SUBTITLES_BEFORE_OVERWRITING_WARNING[LANGUAGE]
-)
-PROMPT_ADDITIONAL_FFSUBSYNC_ARGS = texts.PROMPT_ADDITIONAL_FFSUBSYNC_ARGS[LANGUAGE]
-PROMPT_ADDITIONAL_ALASS_ARGS = texts.PROMPT_ADDITIONAL_ALASS_ARGS[LANGUAGE]
-LABEL_ADDITIONAL_FFSUBSYNC_ARGS = texts.LABEL_ADDITIONAL_FFSUBSYNC_ARGS[LANGUAGE]
-LABEL_ADDITIONAL_ALASS_ARGS = texts.LABEL_ADDITIONAL_ALASS_ARGS[LANGUAGE]
-LABEL_CHECK_VIDEO_FOR_SUBTITLE_STREAM = texts.LABEL_CHECK_VIDEO_FOR_SUBTITLE_STREAM[
-    LANGUAGE
-]
-LABEL_ADD_PREFIX = texts.LABEL_ADD_PREFIX[LANGUAGE]
-LABEL_BACKUP_SUBTITLES = texts.LABEL_BACKUP_SUBTITLES[LANGUAGE]
-LABEL_KEEP_CONVERTED_SUBTITLES = texts.LABEL_KEEP_CONVERTED_SUBTITLES[LANGUAGE]
-LABEL_KEEP_EXTRACTED_SUBTITLES = texts.LABEL_KEEP_EXTRACTED_SUBTITLES[LANGUAGE]
-LABEL_REMEMBER_THE_CHANGES = texts.LABEL_REMEMBER_THE_CHANGES[LANGUAGE]
-LABEL_RESET_TO_DEFAULT_SETTINGS = texts.LABEL_RESET_TO_DEFAULT_SETTINGS[LANGUAGE]
-# Options
-DEFAULT = texts.DEFAULT[LANGUAGE]
-VOICE_ACTIVITY_DETECTOR = texts.VOICE_ACTIVITY_DETECTOR[LANGUAGE]
-LABEL_KEEP_LOG_RECORDS = texts.LABEL_KEEP_LOG_RECORDS[LANGUAGE]
-LABEL_OPEN_LOGS_FOLDER = texts.LABEL_OPEN_LOGS_FOLDER[LANGUAGE]
-LABEL_CLEAR_ALL_LOGS = texts.LABEL_CLEAR_ALL_LOGS[LANGUAGE]
-LOG_FILES_DELETE_WARNING = texts.LOG_FILES_DELETE_WARNING[LANGUAGE]
-SYNC_TOOL_FFSUBSYNC = texts.SYNC_TOOL_FFSUBSYNC[LANGUAGE]
-SYNC_TOOL_ALASS = texts.SYNC_TOOL_ALASS[LANGUAGE]
-OPTION_SAVE_NEXT_TO_SUBTITLE = texts.OPTION_SAVE_NEXT_TO_SUBTITLE[LANGUAGE]
-OPTION_SAVE_NEXT_TO_VIDEO = texts.OPTION_SAVE_NEXT_TO_VIDEO[LANGUAGE]
-OPTION_SAVE_NEXT_TO_VIDEO_WITH_SAME_FILENAME = (
-    texts.OPTION_SAVE_NEXT_TO_VIDEO_WITH_SAME_FILENAME[LANGUAGE]
-)
-OPTION_SAVE_TO_DESKTOP = texts.OPTION_SAVE_TO_DESKTOP[LANGUAGE]
-OPTION_REPLACE_ORIGINAL_SUBTITLE = texts.OPTION_REPLACE_ORIGINAL_SUBTITLE[LANGUAGE]
-OPTION_SELECT_DESTINATION_FOLDER = texts.OPTION_SELECT_DESTINATION_FOLDER[LANGUAGE]
-CHECKBOX_NO_FIX_FRAMERATE = texts.CHECKBOX_NO_FIX_FRAMERATE[LANGUAGE]
-CHECKBOX_GSS = texts.CHECKBOX_GSS[LANGUAGE]
-CHECKBOX_VAD = texts.CHECKBOX_VAD[LANGUAGE]
-LABEL_SPLIT_PENALTY = texts.LABEL_SPLIT_PENALTY[LANGUAGE]
-PAIR_FILES_TITLE = texts.PAIR_FILES_TITLE[LANGUAGE]
-PAIR_FILES_MESSAGE = texts.PAIR_FILES_MESSAGE[LANGUAGE]
-UNPAIRED_SUBTITLES_TITLE = texts.UNPAIRED_SUBTITLES_TITLE[LANGUAGE]
-UNPAIRED_SUBTITLES_MESSAGE = texts.UNPAIRED_SUBTITLES_MESSAGE[LANGUAGE]
-NO_VIDEO = texts.NO_VIDEO[LANGUAGE]
-NO_SUBTITLE = texts.NO_SUBTITLE[LANGUAGE]
-VIDEO_OR_SUBTITLE_TEXT = texts.VIDEO_OR_SUBTITLE_TEXT[LANGUAGE]
-VIDEO_INPUT_LABEL = texts.VIDEO_INPUT_LABEL[LANGUAGE]
-SUBTITLE_INPUT_LABEL = texts.SUBTITLE_INPUT_LABEL[LANGUAGE]
-SUBTITLE_FILES_TEXT = texts.SUBTITLE_FILES_TEXT[LANGUAGE]
-CONTEXT_MENU_REMOVE = texts.CONTEXT_MENU_REMOVE[LANGUAGE]
-CONTEXT_MENU_CHANGE = texts.CONTEXT_MENU_CHANGE[LANGUAGE]
-CONTEXT_MENU_ADD_PAIR = texts.CONTEXT_MENU_ADD_PAIR[LANGUAGE]
-CONTEXT_MENU_ADD_PAIRS = texts.CONTEXT_MENU_ADD_PAIRS[LANGUAGE]
-CONTEXT_MENU_CLEAR_ALL = texts.CONTEXT_MENU_CLEAR_ALL[LANGUAGE]
-CONTEXT_MENU_SHOW_PATH = texts.CONTEXT_MENU_SHOW_PATH[LANGUAGE]
-CONTEXT_MENU_COPY = texts.CONTEXT_MENU_COPY[LANGUAGE]
-BUTTON_ADD_FILES = texts.BUTTON_ADD_FILES[LANGUAGE]
-MENU_ADD_FOLDER = texts.MENU_ADD_FOLDER[LANGUAGE]
-MENU_ADD_MULTIPLE_FILES = texts.MENU_ADD_MULTIPLE_FILES[LANGUAGE]
-MENU_ADD_REFERENCE_SUBITLE_SUBTITLE_PAIRIS = (
-    texts.MENU_ADD_REFERENCE_SUBITLE_SUBTITLE_PAIRIS[LANGUAGE]
-)
-ALASS_SPEED_OPTIMIZATION_TEXT = texts.ALASS_SPEED_OPTIMIZATION_TEXT[LANGUAGE]
-ALASS_DISABLE_FPS_GUESSING_TEXT = texts.ALASS_DISABLE_FPS_GUESSING_TEXT[LANGUAGE]
-REF_DROP_TEXT = texts.REF_DROP_TEXT[LANGUAGE]
-SUB_DROP_TEXT = texts.SUB_DROP_TEXT[LANGUAGE]
-REF_LABEL_TEXT = texts.REF_LABEL_TEXT[LANGUAGE]
-SUB_LABEL_TEXT = texts.SUB_LABEL_TEXT[LANGUAGE]
-PROCESS_PAIRS = texts.PROCESS_PAIRS[LANGUAGE]
-SYNC_TOOL_LABEL_TEXT = texts.SYNC_TOOL_LABEL_TEXT[LANGUAGE]
-EXPLANATION_TEXT_IN_REFERENCE_SUBTITLE_PAIRING = (
-    texts.EXPLANATION_TEXT_IN_REFERENCE_SUBTITLE_PAIRING[LANGUAGE]
-)
-THEME_TEXT = texts.THEME_TEXT[LANGUAGE]
-THEME_SYSTEM_TEXT = texts.THEME_SYSTEM_TEXT[LANGUAGE]
-THEME_DARK_TEXT = texts.THEME_DARK_TEXT[LANGUAGE]
-THEME_LIGHT_TEXT = texts.THEME_LIGHT_TEXT[LANGUAGE]
-# Log messages
-SUCCESS_LOG_TEXT = texts.SUCCESS_LOG_TEXT[LANGUAGE]
-SYNC_SUCCESS_MESSAGE = texts.SYNC_SUCCESS_MESSAGE[LANGUAGE]
-ERROR_SAVING_SUBTITLE = texts.ERROR_SAVING_SUBTITLE[LANGUAGE]
-NON_ZERO_MILLISECONDS = texts.NON_ZERO_MILLISECONDS[LANGUAGE]
-SELECT_ONLY_ONE_OPTION = texts.SELECT_ONLY_ONE_OPTION[LANGUAGE]
-VALID_NUMBER_MILLISECONDS = texts.VALID_NUMBER_MILLISECONDS[LANGUAGE]
-DIALOG_TITLE_TEXT = texts.DIALOG_TITLE_TEXT[LANGUAGE]
-FONT_FAMILY_LABEL_TEXT = texts.FONT_FAMILY_LABEL_TEXT[LANGUAGE]
-FONT_SIZE_LABEL_TEXT = texts.FONT_SIZE_LABEL_TEXT[LANGUAGE]
-FONT_STYLE_LABEL_TEXT = texts.FONT_STYLE_LABEL_TEXT[LANGUAGE]
-BOLD_TEXT = texts.BOLD_TEXT[LANGUAGE]
-ITALIC_TEXT = texts.ITALIC_TEXT[LANGUAGE]
-UNDERLINE_TEXT = texts.UNDERLINE_TEXT[LANGUAGE]
-STRIKETHROUGH_TEXT = texts.STRIKETHROUGH_TEXT[LANGUAGE]
-FONT_INFORMATION_TEXT = texts.FONT_INFORMATION_TEXT[LANGUAGE]
-APPLY_TEXT = texts.APPLY_TEXT[LANGUAGE]
-SELECT_SUBTITLE = texts.SELECT_SUBTITLE[LANGUAGE]
-SELECT_VIDEO = texts.SELECT_VIDEO[LANGUAGE]
-SELECT_VIDEO_OR_SUBTITLE = texts.SELECT_VIDEO_OR_SUBTITLE[LANGUAGE]
-DROP_VIDEO_SUBTITLE_PAIR = texts.DROP_VIDEO_SUBTITLE_PAIR[LANGUAGE]
-DROP_VIDEO_OR_SUBTITLE = texts.DROP_VIDEO_OR_SUBTITLE[LANGUAGE]
-DROP_SUBTITLE_FILE = texts.DROP_SUBTITLE_FILE[LANGUAGE]
-DROP_SINGLE_SUBTITLE_FILE = texts.DROP_SINGLE_SUBTITLE_FILE[LANGUAGE]
-DROP_SINGLE_SUBTITLE_PAIR = texts.DROP_SINGLE_SUBTITLE_PAIR[LANGUAGE]
-SELECT_BOTH_FILES = texts.SELECT_BOTH_FILES[LANGUAGE]
-SELECT_DIFFERENT_FILES = texts.SELECT_DIFFERENT_FILES[LANGUAGE]
-SUBTITLE_FILE_NOT_EXIST = texts.SUBTITLE_FILE_NOT_EXIST[LANGUAGE]
-VIDEO_FILE_NOT_EXIST = texts.VIDEO_FILE_NOT_EXIST[LANGUAGE]
-ERROR_LOADING_SUBTITLE = texts.ERROR_LOADING_SUBTITLE[LANGUAGE]
-ERROR_CONVERT_TIMESTAMP = texts.ERROR_CONVERT_TIMESTAMP[LANGUAGE]
-ERROR_PARSING_TIME_STRING = texts.ERROR_PARSING_TIME_STRING[LANGUAGE]
-ERROR_PARSING_TIME_STRING_DETAILED = texts.ERROR_PARSING_TIME_STRING_DETAILED[LANGUAGE]
-NO_FILES_TO_SYNC = texts.NO_FILES_TO_SYNC[LANGUAGE]
-NO_VALID_FILE_PAIRS = texts.NO_VALID_FILE_PAIRS[LANGUAGE]
-ERROR_PROCESS_TERMINATION = texts.ERROR_PROCESS_TERMINATION[LANGUAGE]
-BATCH_SYNC_CANCEL_CONFIRMATION = texts.BATCH_SYNC_CANCEL_CONFIRMATION[LANGUAGE]
-AUTO_SYNC_CANCELLED = texts.AUTO_SYNC_CANCELLED[LANGUAGE]
-BATCH_SYNC_CANCELLED = texts.BATCH_SYNC_CANCELLED[LANGUAGE]
-NO_SYNC_PROCESS = texts.NO_SYNC_PROCESS[LANGUAGE]
-INVALID_SYNC_TOOL = texts.INVALID_SYNC_TOOL[LANGUAGE]
-FAILED_START_PROCESS = texts.FAILED_START_PROCESS[LANGUAGE]
-ERROR_OCCURRED = texts.ERROR_OCCURRED[LANGUAGE]
-ERROR_DECODING_SUBTITLE = texts.ERROR_DECODING_SUBTITLE[LANGUAGE]
-ERROR_EXECUTING_COMMAND = texts.ERROR_EXECUTING_COMMAND[LANGUAGE]
-DROP_VALID_FILES = texts.DROP_VALID_FILES[LANGUAGE]
-PAIRS_ADDED = texts.PAIRS_ADDED[LANGUAGE]
-UNPAIRED_FILES_ADDED = texts.UNPAIRED_FILES_ADDED[LANGUAGE]
-UNPAIRED_FILES = texts.UNPAIRED_FILES[LANGUAGE]
-DUPLICATE_PAIRS_SKIPPED = texts.DUPLICATE_PAIRS_SKIPPED[LANGUAGE]
-PAIR_ALREADY_EXISTS = texts.PAIR_ALREADY_EXISTS[LANGUAGE]
-PAIR_ADDED = texts.PAIR_ADDED[LANGUAGE]
-SAME_FILE_ERROR = texts.SAME_FILE_ERROR[LANGUAGE]
-PAIR_ALREADY_EXISTS = texts.PAIR_ALREADY_EXISTS[LANGUAGE]
-SUBTITLE_ADDED = texts.SUBTITLE_ADDED[LANGUAGE]
-VIDEO_ADDED = texts.VIDEO_ADDED[LANGUAGE]
-FILE_CHANGED = texts.FILE_CHANGED[LANGUAGE]
-SELECT_ITEM_TO_CHANGE = texts.SELECT_ITEM_TO_CHANGE[LANGUAGE]
-SELECT_ITEM_TO_REMOVE = texts.SELECT_ITEM_TO_REMOVE[LANGUAGE]
-FILE_NOT_EXIST = texts.FILE_NOT_EXIST[LANGUAGE]
-MULTIPLE_ARGUMENTS = texts.MULTIPLE_ARGUMENTS[LANGUAGE]
-INVALID_FILE_FORMAT = texts.INVALID_FILE_FORMAT[LANGUAGE]
-INVALID_SYNC_TOOL_SELECTED = texts.INVALID_SYNC_TOOL_SELECTED[LANGUAGE]
-TEXT_SELECTED_FOLDER = texts.TEXT_SELECTED_FOLDER[LANGUAGE]
-TEXT_NO_FOLDER_SELECTED = texts.TEXT_NO_FOLDER_SELECTED[LANGUAGE]
-TEXT_DESTINATION_FOLDER_DOES_NOT_EXIST = texts.TEXT_DESTINATION_FOLDER_DOES_NOT_EXIST[
-    LANGUAGE
-]
-ADDED_PAIRS_MSG = texts.ADDED_PAIRS_MSG[LANGUAGE]
-SKIPPED_DUPLICATES_MSG = texts.SKIPPED_DUPLICATES_MSG[LANGUAGE]
-NO_VALID_SUBTITLE_PAIRS_TO_PROCESS = texts.NO_VALID_SUBTITLE_PAIRS_TO_PROCESS[LANGUAGE]
-NO_MATCHING_SUBTITLE_PAIRS_FOUND = texts.NO_MATCHING_SUBTITLE_PAIRS_FOUND[LANGUAGE]
-NO_SUBTITLE_PAIRS_TO_PROCESS = texts.NO_SUBTITLE_PAIRS_TO_PROCESS[LANGUAGE]
-NOT_FIND_COMPATIBLE_FILE_MANAGER = texts.NOT_FIND_COMPATIBLE_FILE_MANAGER[LANGUAGE]
-FILE_NOT_FOUND = texts.FILE_NOT_FOUND[LANGUAGE]
-ERROR_OPENING_DIRECTORY = texts.ERROR_OPENING_DIRECTORY[LANGUAGE]
-CONFIG_FILE_NOT_FOUND = texts.CONFIG_FILE_NOT_FOUND[LANGUAGE]
-# Log window messages
-VIDEO_FILE_NOT_FOUND = texts.VIDEO_FILE_NOT_FOUND[LANGUAGE]
-FFPROBE_FAILED = texts.FFPROBE_FAILED[LANGUAGE]
-SUBTITLE_EXTRACTION_FAILED = texts.SUBTITLE_EXTRACTION_FAILED[LANGUAGE]
-SYNCING_LOG_TEXT = texts.SYNCING_LOG_TEXT[LANGUAGE]
-INVALID_PARENT_ITEM = texts.INVALID_PARENT_ITEM[LANGUAGE]
-SKIP_NO_VIDEO_NO_SUBTITLE = texts.SKIP_NO_VIDEO_NO_SUBTITLE[LANGUAGE]
-SKIP_NO_SUBTITLE = texts.SKIP_NO_SUBTITLE[LANGUAGE]
-SKIP_NO_VIDEO = texts.SKIP_NO_VIDEO[LANGUAGE]
-SKIP_UNPAIRED_ITEM = texts.SKIP_UNPAIRED_ITEM[LANGUAGE]
-SKIPPING_ALREADY_SYNCED = texts.SKIPPING_ALREADY_SYNCED[LANGUAGE]
-CONVERTING_SUBTITLE = texts.CONVERTING_SUBTITLE[LANGUAGE]
-ERROR_CONVERTING_SUBTITLE = texts.ERROR_CONVERTING_SUBTITLE[LANGUAGE]
-SUBTITLE_CONVERTED = texts.SUBTITLE_CONVERTED[LANGUAGE]
-ERROR_UNSUPPORTED_CONVERSION = texts.ERROR_UNSUPPORTED_CONVERSION[LANGUAGE]
-FAILED_CONVERT_SUBTITLE = texts.FAILED_CONVERT_SUBTITLE[LANGUAGE]
-FAILED_CONVERT_VIDEO = texts.FAILED_CONVERT_VIDEO[LANGUAGE]
-SPLIT_PENALTY_ZERO = texts.SPLIT_PENALTY_ZERO[LANGUAGE]
-SPLIT_PENALTY_SET = texts.SPLIT_PENALTY_SET[LANGUAGE]
-USING_REFERENCE_SUBTITLE = texts.USING_REFERENCE_SUBTITLE[LANGUAGE]
-USING_VIDEO_FOR_SYNC = texts.USING_VIDEO_FOR_SYNC[LANGUAGE]
-ENABLED_NO_FIX_FRAMERATE = texts.ENABLED_NO_FIX_FRAMERATE[LANGUAGE]
-ENABLED_GSS = texts.ENABLED_GSS[LANGUAGE]
-ADDITIONAL_ARGS_ADDED = texts.ADDITIONAL_ARGS_ADDED[LANGUAGE]
-SYNCING_STARTED = texts.SYNCING_STARTED[LANGUAGE]
-SYNCING_ENDED = texts.SYNCING_ENDED[LANGUAGE]
-SYNC_SUCCESS = texts.SYNC_SUCCESS[LANGUAGE]
-SYNC_ERROR = texts.SYNC_ERROR[LANGUAGE]
-SYNC_ERROR_OCCURRED = texts.SYNC_ERROR_OCCURRED[LANGUAGE]
-BATCH_SYNC_COMPLETED = texts.BATCH_SYNC_COMPLETED[LANGUAGE]
-PREPARING_SYNC = texts.PREPARING_SYNC[LANGUAGE]
-CONVERTING_TTML = texts.CONVERTING_TTML[LANGUAGE]
-CONVERTING_VTT = texts.CONVERTING_VTT[LANGUAGE]
-CONVERTING_SBV = texts.CONVERTING_SBV[LANGUAGE]
-CONVERTING_SUB = texts.CONVERTING_SUB[LANGUAGE]
-CONVERTING_ASS = texts.CONVERTING_ASS[LANGUAGE]
-CONVERTING_STL = texts.CONVERTING_STL[LANGUAGE]
-ERROR_CONVERTING_SUBTITLE = texts.ERROR_CONVERTING_SUBTITLE[LANGUAGE]
-CONVERSION_NOT_SUPPORTED = texts.CONVERSION_NOT_SUPPORTED[LANGUAGE]
-SUBTITLE_CONVERTED = texts.SUBTITLE_CONVERTED[LANGUAGE]
-RETRY_ENCODING_MSG = texts.RETRY_ENCODING_MSG[LANGUAGE]
-ALASS_SPEED_OPTIMIZATION_LOG = texts.ALASS_SPEED_OPTIMIZATION_LOG[LANGUAGE]
-ALASS_DISABLE_FPS_GUESSING_LOG = texts.ALASS_DISABLE_FPS_GUESSING_LOG[LANGUAGE]
-CHANGING_ENCODING_MSG = texts.CHANGING_ENCODING_MSG[LANGUAGE]
-ENCODING_CHANGED_MSG = texts.ENCODING_CHANGED_MSG[LANGUAGE]
-SYNC_SUCCESS_COUNT = texts.SYNC_SUCCESS_COUNT[LANGUAGE]
-SYNC_FAILURE_COUNT = texts.SYNC_FAILURE_COUNT[LANGUAGE]
-SYNC_FAILURE_COUNT_BATCH = texts.SYNC_FAILURE_COUNT_BATCH[LANGUAGE]
-ERROR_CHANGING_ENCODING_MSG = texts.ERROR_CHANGING_ENCODING_MSG[LANGUAGE]
-BACKUP_CREATED = texts.BACKUP_CREATED[LANGUAGE]
-CHECKING_SUBTITLE_STREAMS = texts.CHECKING_SUBTITLE_STREAMS[LANGUAGE]
-FOUND_COMPATIBLE_SUBTITLES = texts.FOUND_COMPATIBLE_SUBTITLES[LANGUAGE]
-NO_COMPATIBLE_SUBTITLES = texts.NO_COMPATIBLE_SUBTITLES[LANGUAGE]
-SUCCESSFULLY_EXTRACTED = texts.SUCCESSFULLY_EXTRACTED[LANGUAGE]
-CHOOSING_BEST_SUBTITLE = texts.CHOOSING_BEST_SUBTITLE[LANGUAGE]
-CHOSEN_SUBTITLE = texts.CHOSEN_SUBTITLE[LANGUAGE]
-FAILED_TO_EXTRACT_SUBTITLES = texts.FAILED_TO_EXTRACT_SUBTITLES[LANGUAGE]
-USED_THE_LONGEST_SUBTITLE = texts.USED_THE_LONGEST_SUBTITLE[LANGUAGE]
-DELETING_EXTRACTED_SUBTITLE_FOLDER = texts.DELETING_EXTRACTED_SUBTITLE_FOLDER[LANGUAGE]
-DELETING_CONVERTED_SUBTITLE = texts.DELETING_CONVERTED_SUBTITLE[LANGUAGE]
-ADDED_FILES_TEXT = texts.ADDED_FILES_TEXT[LANGUAGE]
-SKIPPED_DUPLICATE_FILES_TEXT = texts.SKIPPED_DUPLICATE_FILES_TEXT[LANGUAGE]
-SKIPPED_OTHER_LIST_FILES_TEXT = texts.SKIPPED_OTHER_LIST_FILES_TEXT[LANGUAGE]
-SKIPPED_SEASON_EPISODE_DUPLICATES_TEXT = texts.SKIPPED_SEASON_EPISODE_DUPLICATES_TEXT[
-    LANGUAGE
-]
-SKIPPED_INVALID_FORMAT_FILES_TEXT = texts.SKIPPED_INVALID_FORMAT_FILES_TEXT[LANGUAGE]
-NO_FILES_SELECTED = texts.NO_FILES_SELECTED[LANGUAGE]
-NO_ITEM_SELECTED_TO_REMOVE = texts.NO_ITEM_SELECTED_TO_REMOVE[LANGUAGE]
-NO_FILES_SELECTED_TO_SHOW_PATH = texts.NO_FILES_SELECTED_TO_SHOW_PATH[LANGUAGE]
-REMOVED_ITEM = texts.REMOVED_ITEM[LANGUAGE]
-FILES_MUST_CONTAIN_PATTERNS = texts.FILES_MUST_CONTAIN_PATTERNS[LANGUAGE]
-NO_VALID_SUBTITLE_FILES = texts.NO_VALID_SUBTITLE_FILES[LANGUAGE]
-default_encoding = sys.getfilesystemencoding()
-
-
-def create_process(cmd):
-    kwargs = {
-        "shell": True,
-        "stdout": subprocess.PIPE,
-        "stderr": subprocess.STDOUT,
-        "universal_newlines": True,
-        "encoding": default_encoding,
-        "errors": "replace",
-    }
-
-    if platform == "Windows":
-        startupinfo = subprocess.STARTUPINFO()
-        startupinfo.dwFlags |= subprocess.STARTF_USESHOWWINDOW
-        startupinfo.wShowWindow = subprocess.SW_HIDE
-        kwargs.update(
-            {"startupinfo": startupinfo, "creationflags": subprocess.CREATE_NO_WINDOW}
-        )
-
-    return subprocess.Popen(cmd, **kwargs)
-
-
-def update_config(key, value):
-    if remember_the_changes or key == "remember_the_changes":
-        try:
-            with open(config_path, "r", encoding="utf-8") as config_file:
-                config = json.load(config_file)
-        except FileNotFoundError:
-            config = {}
-            messagebox.showerror("Error", CONFIG_FILE_NOT_FOUND)
-        config[key] = value
-        with open(config_path, "w", encoding="utf-8") as config_file:
-            json.dump(config, config_file, indent=4)
-
-
-def detect_encoding(file_path):
-    with open(file_path, "rb") as f:
-        raw_data = f.read()
-    detected_encoding = None
-    for detectors in (cchardet, charset_normalizer, chardet):
-        try:
-            result = detectors.detect(raw_data)["encoding"]
-        except Exception:
-            continue
-        if result is not None:
-            detected_encoding = result
-            break
-    encoding = detected_encoding if detected_encoding else "utf-8"
-    return encoding.lower()
-
-
 # Shift Subtitle Start
-total_shifted_milliseconds = {}
 
+total_shifted_milliseconds = {}
 
 def shift_subtitle(subtitle_file, milliseconds, save_to_desktop, replace_original):
     # Load file with encoding detection using detect_encoding function
@@ -956,7 +380,6 @@ def shift_subtitle(subtitle_file, milliseconds, save_to_desktop, replace_origina
             if not response:
                 return
     elif save_to_desktop:
-        desktop_path = get_desktop_path()
         new_subtitle_file = os.path.join(desktop_path, f"{total_shifted}ms_{filename}")
     else:
         new_subtitle_file = os.path.join(
@@ -1184,101 +607,12 @@ def place_window_top_right(event=None, margin=50):
     root.geometry(f"{width}x{height}+{x}+{y}")
 
 
-class ToolTip:
-    def __init__(self, widget, text):
-        self.widget = widget
-        self.text = text
-        self.tooltip = None
-        self.widget.bind("<Enter>", self.show_tooltip)
-        self.widget.bind("<Leave>", self.hide_tooltip)
-        self.widget.bind("<Button-1>", self.hide_tooltip)
-
-    def show_tooltip(self, event=None):
-        if self.tooltip is None:
-            # Get the position of the widget
-            x_pos = self.widget.winfo_rootx()
-            y_pos = (
-                self.widget.winfo_rooty() + self.widget.winfo_height()
-            )  # Adjust tooltip position below the widget
-            # Calculate the screen dimensions
-            screen_width = self.widget.winfo_screenwidth()
-            screen_height = self.widget.winfo_screenheight()
-            # Create a temporary label to calculate the width based on content
-            temp_label = tk.Label(text=self.text)
-            # removed font=("tahoma", "8", "normal") from label
-            temp_label.update_idletasks()
-            content_width = (
-                temp_label.winfo_reqwidth()
-            )  # Get the required width of the content
-            # Set the tooltip width based on content width, limited to a maximum of 200
-            tooltip_width = min(content_width, 200)
-            # Calculate wraplength dynamically
-            wraplength = min(content_width, 200)
-            # Create the tooltip at the calculated position
-            self.tooltip = tk.Toplevel(self.widget)
-            self.tooltip.wm_overrideredirect(True)
-            self.tooltip.attributes("-topmost", True)  # Make the tooltip window topmost
-            # Adjust tooltip position to stay within the screen bounds
-            if x_pos + tooltip_width > screen_width:
-                x_pos = screen_width - tooltip_width
-            if y_pos + self.widget.winfo_height() > screen_height:
-                y_pos = screen_height - self.widget.winfo_height() - 3
-            # Adjust tooltip position to avoid covering the button
-            y_pos = max(y_pos, 0)
-            # Adjust tooltip position if too far to the left
-            x_pos = max(x_pos, 0)
-            self.tooltip.wm_geometry("+%d+%d" % (x_pos, y_pos))
-            label = tk.Label(
-                self.tooltip,
-                text=self.text,
-                justify=tk.LEFT,
-                wraplength=wraplength,
-                background="#ffffe0",
-                foreground="black",
-                relief=tk.SOLID,
-                borderwidth=1,
-            )
-            # removed font=("tahoma", "8", "normal") from label
-            label.pack(ipadx=1)
-
-    def hide_tooltip(self, event=None):
-        if self.tooltip:
-            self.tooltip.destroy()
-            self.tooltip = None
-
-
 def on_manual_tab_selected(event=None):
     # Set focus to label_drop_box to prevent autofocus on entry_milliseconds
     label_drop_box.focus_set()
     # Insert default value "0" into entry_milliseconds if it's empty
     if not entry_milliseconds.get():
         entry_milliseconds.insert(0, "0")
-
-
-def dark_title_bar(window):
-    if platform == "Windows" and THEME == "dark":
-        try:
-            window.update()
-            DWMWA_USE_IMMERSIVE_DARK_MODE = 20
-            set_window_attribute = ctypes.windll.dwmapi.DwmSetWindowAttribute
-            get_parent = ctypes.windll.user32.GetParent
-            hwnd = get_parent(window.winfo_id())
-            rendering_policy = DWMWA_USE_IMMERSIVE_DARK_MODE
-            value = 2
-            value = ctypes.c_int(value)
-            set_window_attribute(
-                hwnd, rendering_policy, ctypes.byref(value), ctypes.sizeof(value)
-            )
-        except Exception:
-            pass
-    elif platform == "Linux":
-        # No direct equivalent for Linux, but you could set window properties
-        try:
-            window.attributes("-type", "normal")
-            if THEME == "dark":
-                window.attributes("-alpha", 0.95)  # Slightly transparent for dark theme
-        except Exception:
-            pass
 
 
 root = TkinterDnD.Tk()
@@ -1676,24 +1010,6 @@ def change_log_window_font():
         pass
 
 
-def restart_program():
-    python = sys.executable
-    if getattr(sys, "frozen", False):
-        # Running as an executable
-        os.execl(python, python)
-    else:
-        # Running as a script
-        script_path = os.path.abspath(__file__)
-        os.execl(python, python, script_path)
-
-
-def reset_to_default_settings():
-    if messagebox.askyesno(WARNING, CONFIRM_RESET_MESSAGE):
-        for key, value in default_settings.items():
-            update_config(key, value)
-        restart_program()
-
-
 def toggle_keep_converted_subtitles():
     global keep_converted_subtitles
     if keep_converted_subtitles:
@@ -1819,7 +1135,7 @@ def set_theme(theme):
 
 
 def open_logs_folder():
-    logs_folder = os.path.join(base_dir, f"{PROGRAM_NAME}_logs")
+    logs_folder = os.path.join(config_dir, f"{PROGRAM_NAME}_logs")
     # Ensure logs directory exists
     os.makedirs(logs_folder, exist_ok=True)
 
@@ -1856,7 +1172,7 @@ def open_logs_folder():
 
 def check_logs_exist():
     try:
-        logs_folder = os.path.join(base_dir, f"{PROGRAM_NAME}_logs")
+        logs_folder = os.path.join(config_dir, f"{PROGRAM_NAME}_logs")
         txt_files = [
             f
             for f in os.listdir(logs_folder)
@@ -1868,7 +1184,7 @@ def check_logs_exist():
 
 
 def clear_all_logs():
-    logs_folder = os.path.join(base_dir, f"{PROGRAM_NAME}_logs")
+    logs_folder = os.path.join(config_dir, f"{PROGRAM_NAME}_logs")
     if os.path.exists(logs_folder):
         num_txt_files = check_logs_exist()
         if num_txt_files == 0:
@@ -2599,7 +1915,7 @@ def save_log_file(log_window, suffix=""):
     if keep_logs:
         # Save log window content to a log file
         log_content = log_window.get("1.0", tk.END)
-        logs_folder = os.path.join(base_dir, f"{PROGRAM_NAME}_logs")
+        logs_folder = os.path.join(config_dir, f"{PROGRAM_NAME}_logs")
         os.makedirs(logs_folder, exist_ok=True)
         timestamp = datetime.now().strftime("%Y-%m-%d_%H-%M-%S")
         log_filename = os.path.join(logs_folder, f"{timestamp}{suffix}.txt")
@@ -2663,28 +1979,6 @@ def kill_process_tree(pid):
                     pass
     except psutil.NoSuchProcess:
         pass
-
-
-def get_desktop_path():
-    if platform == "Windows":
-        try:
-            import winreg
-
-            reg_key = winreg.OpenKey(
-                winreg.HKEY_CURRENT_USER,
-                r"Software\Microsoft\Windows\CurrentVersion\Explorer\Shell Folders",
-            )
-            desktop = winreg.QueryValueEx(reg_key, "Desktop")[0]
-            winreg.CloseKey(reg_key)
-        except Exception:
-            desktop = "None"
-    else:
-        from pathlib import Path
-
-        desktop = Path.home() / "Desktop"
-    if not os.path.exists(desktop):
-        desktop = os.path.normpath(os.path.expanduser("~"))
-    return desktop
 
 
 def levenshtein_distance(s1, s2):
@@ -2954,7 +2248,6 @@ def start_batch_sync():
                     continue
                 # Prepare output file path
                 if action_var_auto.get() == OPTION_SAVE_TO_DESKTOP:
-                    desktop_path = get_desktop_path()
                     base_output_dir = desktop_path
                 elif action_var_auto.get() == OPTION_REPLACE_ORIGINAL_SUBTITLE:
                     base_output_dir = os.path.dirname(subtitle_file)
@@ -3027,7 +2320,6 @@ def start_batch_sync():
 
                 # Determine base output directory for EACH file pair - moved inside the loop
                 if action_var_auto.get() == OPTION_SAVE_TO_DESKTOP:
-                    desktop_path = get_desktop_path()
                     base_output_dir = desktop_path
                 elif action_var_auto.get() == OPTION_REPLACE_ORIGINAL_SUBTITLE:
                     base_output_dir = os.path.dirname(subtitle_file)
@@ -5347,7 +4639,6 @@ def start_automatic_sync():
     if action == OPTION_SAVE_NEXT_TO_VIDEO:
         output_dir = os.path.dirname(video_file)
     elif action == OPTION_SAVE_TO_DESKTOP:
-        desktop_path = get_desktop_path()
         output_dir = desktop_path
     elif action == OPTION_REPLACE_ORIGINAL_SUBTITLE:
         output_subtitle_file = subtitle_file
