+++ main/constants.py
@@ -1,0 +1,59 @@
+# Language selection (ALL TRANSLATIONS ARE LOCATED IN "texts.py")
+LANGUAGES = {
+    "English": "en",
+    "Español": "es",
+    "Deutsch": "de",
+    "Français": "fr",
+    "Italiano": "it",
+    "Polski": "pl",
+    "Português": "pt",
+    "Türkçe": "tr",
+    "Tiếng Việt": "vi",
+    "Bahasa Indonesia": "id",
+    "Bahasa Melayu": "ms",
+    "ไทย": "th",
+    "Українська": "uk",
+    "Русский": "ru",
+    "中国人": "zh",
+    "日本語": "ja",
+    "한국어": "ko",
+    "हिन्दी": "hi",
+    "বাংলা": "bn",
+    "اردو": "ur",
+    "العربية": "ar",
+    "فارسی": "fa",
+}
+
+# File extensions
+FFSUBSYNC_SUPPORTED_EXTENSIONS = [".srt", ".ass", ".ssa", ".vtt"]
+ALASS_SUPPORTED_EXTENSIONS = [".srt", ".ass", ".ssa", ".sub", ".idx"]
+ALASS_EXTRACTABLE_SUBTITLE_EXTENSIONS = {"subrip": "srt", "ass": "ass", "webvtt": "vtt"}
+SUBTITLE_EXTENSIONS = [
+    ".srt",
+    ".vtt",
+    ".sbv",
+    ".sub",
+    ".ass",
+    ".ssa",
+    ".dfxp",
+    ".ttml",
+    ".itt",
+    ".stl",
+    ".idx",
+]
+VIDEO_EXTENSIONS = [
+    ".mp4",
+    ".mkv",
+    ".avi",
+    ".webm",
+    ".flv",
+    ".mov",
+    ".wmv",
+    ".mpg",
+    ".mpeg",
+    ".m4v",
+    ".3gp",
+    ".h264",
+    ".h265",
+    ".hevc",
+]
