+++ main/functions/get_platform.py
@@ -1,0 +1,3 @@
+import platform
+
+platform = platform.system()
