+++ main/texts_constants.py
@@ -1,0 +1,279 @@
+import texts
+from config import config
+
+LANGUAGE = config.get("language", "en")
+# Program information
+PROGRAM_NAME = texts.PROGRAM_NAME
+GITHUB_URL = texts.GITHUB_URL
+GITHUB_VERSION_URL = texts.GITHUB_VERSION_URL
+GITHUB_LATEST_RELEASE_URL = texts.GITHUB_LATEST_RELEASE_URL
+# Tooltip texts for checkboxes
+TOOLTIP_SAVE_TO_DESKTOP = texts.TOOLTIP_SAVE_TO_DESKTOP[LANGUAGE]
+TOOLTIP_REPLACE_ORIGINAL = texts.TOOLTIP_REPLACE_ORIGINAL[LANGUAGE]
+TOOLTIP_GSS = texts.TOOLTIP_GSS[LANGUAGE]
+TOOLTIP_VAD = texts.TOOLTIP_VAD[LANGUAGE]
+TOOLTIP_FRAMERATE = texts.TOOLTIP_FRAMERATE[LANGUAGE]
+TOOLTIP_ALASS_SPEED_OPTIMIZATION = texts.TOOLTIP_ALASS_SPEED_OPTIMIZATION[LANGUAGE]
+TOOLTIP_ALASS_DISABLE_FPS_GUESSING = texts.TOOLTIP_ALASS_DISABLE_FPS_GUESSING[LANGUAGE]
+TOOLTIP_TEXT_ACTION_MENU_AUTO = texts.TOOLTIP_TEXT_ACTION_MENU_AUTO[LANGUAGE]
+TOOLTIP_TEXT_SYNC_TOOL_MENU_AUTO = texts.TOOLTIP_TEXT_SYNC_TOOL_MENU_AUTO[LANGUAGE]
+# Dialogs - Buttons
+NOTIFY_ABOUT_UPDATES_TEXT = texts.NOTIFY_ABOUT_UPDATES_TEXT[LANGUAGE]
+UPDATE_AVAILABLE_TEXT = texts.UPDATE_AVAILABLE_TEXT[LANGUAGE]
+UPDATE_AVAILABLE_TITLE = texts.UPDATE_AVAILABLE_TITLE[LANGUAGE]
+LANGUAGE_LABEL_TEXT = texts.LANGUAGE_LABEL_TEXT[LANGUAGE]
+TAB_AUTOMATIC_SYNC = texts.TAB_AUTOMATIC_SYNC[LANGUAGE]
+TAB_MANUAL_SYNC = texts.TAB_MANUAL_SYNC[LANGUAGE]
+CANCEL_TEXT = texts.CANCEL_TEXT[LANGUAGE]
+GENERATE_AGAIN_TEXT = texts.GENERATE_AGAIN_TEXT[LANGUAGE]
+GO_BACK = texts.GO_BACK[LANGUAGE]
+BATCH_MODE_TEXT = texts.BATCH_MODE_TEXT[LANGUAGE]
+NORMAL_MODE_TEXT = texts.NORMAL_MODE_TEXT[LANGUAGE]
+START_AUTOMATIC_SYNC_TEXT = texts.START_AUTOMATIC_SYNC_TEXT[LANGUAGE]
+START_BATCH_SYNC_TEXT = texts.START_BATCH_SYNC_TEXT[LANGUAGE]
+BATCH_INPUT_TEXT = texts.BATCH_INPUT_TEXT[LANGUAGE]
+BATCH_INPUT_LABEL_TEXT = texts.BATCH_INPUT_LABEL_TEXT[LANGUAGE]
+SHIFT_SUBTITLE_TEXT = texts.SHIFT_SUBTITLE_TEXT[LANGUAGE]
+LABEL_SHIFT_SUBTITLE = texts.LABEL_SHIFT_SUBTITLE[LANGUAGE]
+REPLACE_ORIGINAL_TITLE = texts.REPLACE_ORIGINAL_TITLE[LANGUAGE]
+REPLACE_ORIGINAL_TEXT = texts.REPLACE_ORIGINAL_TEXT[LANGUAGE]
+FILE_EXISTS_TITLE = texts.FILE_EXISTS_TITLE[LANGUAGE]
+FILE_EXISTS_TEXT = texts.FILE_EXISTS_TEXT[LANGUAGE]
+ALREADY_SYNCED_FILES_TITLE = texts.ALREADY_SYNCED_FILES_TITLE[LANGUAGE]
+ALREADY_SYNCED_FILES_MESSAGE = texts.ALREADY_SYNCED_FILES_MESSAGE[LANGUAGE]
+SUBTITLE_INPUT_TEXT = texts.SUBTITLE_INPUT_TEXT[LANGUAGE]
+VIDEO_INPUT_TEXT = texts.VIDEO_INPUT_TEXT[LANGUAGE]
+LABEL_DROP_BOX = texts.LABEL_DROP_BOX[LANGUAGE]
+WARNING = texts.WARNING[LANGUAGE]
+CONFIRM_RESET_MESSAGE = texts.CONFIRM_RESET_MESSAGE[LANGUAGE]
+TOGGLE_KEEP_CONVERTED_SUBTITLES_WARNING = texts.TOGGLE_KEEP_CONVERTED_SUBTITLES_WARNING[
+    LANGUAGE
+]
+TOGGLE_KEEP_EXTRACTED_SUBTITLES_WARNING = texts.TOGGLE_KEEP_EXTRACTED_SUBTITLES_WARNING[
+    LANGUAGE
+]
+BACKUP_SUBTITLES_BEFORE_OVERWRITING_WARNING = (
+    texts.BACKUP_SUBTITLES_BEFORE_OVERWRITING_WARNING[LANGUAGE]
+)
+PROMPT_ADDITIONAL_FFSUBSYNC_ARGS = texts.PROMPT_ADDITIONAL_FFSUBSYNC_ARGS[LANGUAGE]
+PROMPT_ADDITIONAL_ALASS_ARGS = texts.PROMPT_ADDITIONAL_ALASS_ARGS[LANGUAGE]
+LABEL_ADDITIONAL_FFSUBSYNC_ARGS = texts.LABEL_ADDITIONAL_FFSUBSYNC_ARGS[LANGUAGE]
+LABEL_ADDITIONAL_ALASS_ARGS = texts.LABEL_ADDITIONAL_ALASS_ARGS[LANGUAGE]
+LABEL_CHECK_VIDEO_FOR_SUBTITLE_STREAM = texts.LABEL_CHECK_VIDEO_FOR_SUBTITLE_STREAM[
+    LANGUAGE
+]
+LABEL_ADD_PREFIX = texts.LABEL_ADD_PREFIX[LANGUAGE]
+LABEL_BACKUP_SUBTITLES = texts.LABEL_BACKUP_SUBTITLES[LANGUAGE]
+LABEL_KEEP_CONVERTED_SUBTITLES = texts.LABEL_KEEP_CONVERTED_SUBTITLES[LANGUAGE]
+LABEL_KEEP_EXTRACTED_SUBTITLES = texts.LABEL_KEEP_EXTRACTED_SUBTITLES[LANGUAGE]
+LABEL_REMEMBER_THE_CHANGES = texts.LABEL_REMEMBER_THE_CHANGES[LANGUAGE]
+LABEL_RESET_TO_DEFAULT_SETTINGS = texts.LABEL_RESET_TO_DEFAULT_SETTINGS[LANGUAGE]
+# Options
+DEFAULT = texts.DEFAULT[LANGUAGE]
+VOICE_ACTIVITY_DETECTOR = texts.VOICE_ACTIVITY_DETECTOR[LANGUAGE]
+LABEL_KEEP_LOG_RECORDS = texts.LABEL_KEEP_LOG_RECORDS[LANGUAGE]
+LABEL_OPEN_LOGS_FOLDER = texts.LABEL_OPEN_LOGS_FOLDER[LANGUAGE]
+LABEL_CLEAR_ALL_LOGS = texts.LABEL_CLEAR_ALL_LOGS[LANGUAGE]
+LOG_FILES_DELETE_WARNING = texts.LOG_FILES_DELETE_WARNING[LANGUAGE]
+SYNC_TOOL_FFSUBSYNC = texts.SYNC_TOOL_FFSUBSYNC[LANGUAGE]
+SYNC_TOOL_ALASS = texts.SYNC_TOOL_ALASS[LANGUAGE]
+OPTION_SAVE_NEXT_TO_SUBTITLE = texts.OPTION_SAVE_NEXT_TO_SUBTITLE[LANGUAGE]
+OPTION_SAVE_NEXT_TO_VIDEO = texts.OPTION_SAVE_NEXT_TO_VIDEO[LANGUAGE]
+OPTION_SAVE_NEXT_TO_VIDEO_WITH_SAME_FILENAME = (
+    texts.OPTION_SAVE_NEXT_TO_VIDEO_WITH_SAME_FILENAME[LANGUAGE]
+)
+OPTION_SAVE_TO_DESKTOP = texts.OPTION_SAVE_TO_DESKTOP[LANGUAGE]
+OPTION_REPLACE_ORIGINAL_SUBTITLE = texts.OPTION_REPLACE_ORIGINAL_SUBTITLE[LANGUAGE]
+OPTION_SELECT_DESTINATION_FOLDER = texts.OPTION_SELECT_DESTINATION_FOLDER[LANGUAGE]
+CHECKBOX_NO_FIX_FRAMERATE = texts.CHECKBOX_NO_FIX_FRAMERATE[LANGUAGE]
+CHECKBOX_GSS = texts.CHECKBOX_GSS[LANGUAGE]
+CHECKBOX_VAD = texts.CHECKBOX_VAD[LANGUAGE]
+LABEL_SPLIT_PENALTY = texts.LABEL_SPLIT_PENALTY[LANGUAGE]
+PAIR_FILES_TITLE = texts.PAIR_FILES_TITLE[LANGUAGE]
+PAIR_FILES_MESSAGE = texts.PAIR_FILES_MESSAGE[LANGUAGE]
+UNPAIRED_SUBTITLES_TITLE = texts.UNPAIRED_SUBTITLES_TITLE[LANGUAGE]
+UNPAIRED_SUBTITLES_MESSAGE = texts.UNPAIRED_SUBTITLES_MESSAGE[LANGUAGE]
+NO_VIDEO = texts.NO_VIDEO[LANGUAGE]
+NO_SUBTITLE = texts.NO_SUBTITLE[LANGUAGE]
+VIDEO_OR_SUBTITLE_TEXT = texts.VIDEO_OR_SUBTITLE_TEXT[LANGUAGE]
+VIDEO_INPUT_LABEL = texts.VIDEO_INPUT_LABEL[LANGUAGE]
+SUBTITLE_INPUT_LABEL = texts.SUBTITLE_INPUT_LABEL[LANGUAGE]
+SUBTITLE_FILES_TEXT = texts.SUBTITLE_FILES_TEXT[LANGUAGE]
+CONTEXT_MENU_REMOVE = texts.CONTEXT_MENU_REMOVE[LANGUAGE]
+CONTEXT_MENU_CHANGE = texts.CONTEXT_MENU_CHANGE[LANGUAGE]
+CONTEXT_MENU_ADD_PAIR = texts.CONTEXT_MENU_ADD_PAIR[LANGUAGE]
+CONTEXT_MENU_ADD_PAIRS = texts.CONTEXT_MENU_ADD_PAIRS[LANGUAGE]
+CONTEXT_MENU_CLEAR_ALL = texts.CONTEXT_MENU_CLEAR_ALL[LANGUAGE]
+CONTEXT_MENU_SHOW_PATH = texts.CONTEXT_MENU_SHOW_PATH[LANGUAGE]
+CONTEXT_MENU_COPY = texts.CONTEXT_MENU_COPY[LANGUAGE]
+BUTTON_ADD_FILES = texts.BUTTON_ADD_FILES[LANGUAGE]
+MENU_ADD_FOLDER = texts.MENU_ADD_FOLDER[LANGUAGE]
+MENU_ADD_MULTIPLE_FILES = texts.MENU_ADD_MULTIPLE_FILES[LANGUAGE]
+MENU_ADD_REFERENCE_SUBITLE_SUBTITLE_PAIRIS = (
+    texts.MENU_ADD_REFERENCE_SUBITLE_SUBTITLE_PAIRIS[LANGUAGE]
+)
+ALASS_SPEED_OPTIMIZATION_TEXT = texts.ALASS_SPEED_OPTIMIZATION_TEXT[LANGUAGE]
+ALASS_DISABLE_FPS_GUESSING_TEXT = texts.ALASS_DISABLE_FPS_GUESSING_TEXT[LANGUAGE]
+REF_DROP_TEXT = texts.REF_DROP_TEXT[LANGUAGE]
+SUB_DROP_TEXT = texts.SUB_DROP_TEXT[LANGUAGE]
+REF_LABEL_TEXT = texts.REF_LABEL_TEXT[LANGUAGE]
+SUB_LABEL_TEXT = texts.SUB_LABEL_TEXT[LANGUAGE]
+PROCESS_PAIRS = texts.PROCESS_PAIRS[LANGUAGE]
+SYNC_TOOL_LABEL_TEXT = texts.SYNC_TOOL_LABEL_TEXT[LANGUAGE]
+EXPLANATION_TEXT_IN_REFERENCE_SUBTITLE_PAIRING = (
+    texts.EXPLANATION_TEXT_IN_REFERENCE_SUBTITLE_PAIRING[LANGUAGE]
+)
+THEME_TEXT = texts.THEME_TEXT[LANGUAGE]
+THEME_SYSTEM_TEXT = texts.THEME_SYSTEM_TEXT[LANGUAGE]
+THEME_DARK_TEXT = texts.THEME_DARK_TEXT[LANGUAGE]
+THEME_LIGHT_TEXT = texts.THEME_LIGHT_TEXT[LANGUAGE]
+# Log messages
+SUCCESS_LOG_TEXT = texts.SUCCESS_LOG_TEXT[LANGUAGE]
+SYNC_SUCCESS_MESSAGE = texts.SYNC_SUCCESS_MESSAGE[LANGUAGE]
+ERROR_SAVING_SUBTITLE = texts.ERROR_SAVING_SUBTITLE[LANGUAGE]
+NON_ZERO_MILLISECONDS = texts.NON_ZERO_MILLISECONDS[LANGUAGE]
+SELECT_ONLY_ONE_OPTION = texts.SELECT_ONLY_ONE_OPTION[LANGUAGE]
+VALID_NUMBER_MILLISECONDS = texts.VALID_NUMBER_MILLISECONDS[LANGUAGE]
+DIALOG_TITLE_TEXT = texts.DIALOG_TITLE_TEXT[LANGUAGE]
+FONT_FAMILY_LABEL_TEXT = texts.FONT_FAMILY_LABEL_TEXT[LANGUAGE]
+FONT_SIZE_LABEL_TEXT = texts.FONT_SIZE_LABEL_TEXT[LANGUAGE]
+FONT_STYLE_LABEL_TEXT = texts.FONT_STYLE_LABEL_TEXT[LANGUAGE]
+BOLD_TEXT = texts.BOLD_TEXT[LANGUAGE]
+ITALIC_TEXT = texts.ITALIC_TEXT[LANGUAGE]
+UNDERLINE_TEXT = texts.UNDERLINE_TEXT[LANGUAGE]
+STRIKETHROUGH_TEXT = texts.STRIKETHROUGH_TEXT[LANGUAGE]
+FONT_INFORMATION_TEXT = texts.FONT_INFORMATION_TEXT[LANGUAGE]
+APPLY_TEXT = texts.APPLY_TEXT[LANGUAGE]
+SELECT_SUBTITLE = texts.SELECT_SUBTITLE[LANGUAGE]
+SELECT_VIDEO = texts.SELECT_VIDEO[LANGUAGE]
+SELECT_VIDEO_OR_SUBTITLE = texts.SELECT_VIDEO_OR_SUBTITLE[LANGUAGE]
+DROP_VIDEO_SUBTITLE_PAIR = texts.DROP_VIDEO_SUBTITLE_PAIR[LANGUAGE]
+DROP_VIDEO_OR_SUBTITLE = texts.DROP_VIDEO_OR_SUBTITLE[LANGUAGE]
+DROP_SUBTITLE_FILE = texts.DROP_SUBTITLE_FILE[LANGUAGE]
+DROP_SINGLE_SUBTITLE_FILE = texts.DROP_SINGLE_SUBTITLE_FILE[LANGUAGE]
+DROP_SINGLE_SUBTITLE_PAIR = texts.DROP_SINGLE_SUBTITLE_PAIR[LANGUAGE]
+SELECT_BOTH_FILES = texts.SELECT_BOTH_FILES[LANGUAGE]
+SELECT_DIFFERENT_FILES = texts.SELECT_DIFFERENT_FILES[LANGUAGE]
+SUBTITLE_FILE_NOT_EXIST = texts.SUBTITLE_FILE_NOT_EXIST[LANGUAGE]
+VIDEO_FILE_NOT_EXIST = texts.VIDEO_FILE_NOT_EXIST[LANGUAGE]
+ERROR_LOADING_SUBTITLE = texts.ERROR_LOADING_SUBTITLE[LANGUAGE]
+ERROR_CONVERT_TIMESTAMP = texts.ERROR_CONVERT_TIMESTAMP[LANGUAGE]
+ERROR_PARSING_TIME_STRING = texts.ERROR_PARSING_TIME_STRING[LANGUAGE]
+ERROR_PARSING_TIME_STRING_DETAILED = texts.ERROR_PARSING_TIME_STRING_DETAILED[LANGUAGE]
+NO_FILES_TO_SYNC = texts.NO_FILES_TO_SYNC[LANGUAGE]
+NO_VALID_FILE_PAIRS = texts.NO_VALID_FILE_PAIRS[LANGUAGE]
+ERROR_PROCESS_TERMINATION = texts.ERROR_PROCESS_TERMINATION[LANGUAGE]
+BATCH_SYNC_CANCEL_CONFIRMATION = texts.BATCH_SYNC_CANCEL_CONFIRMATION[LANGUAGE]
+AUTO_SYNC_CANCELLED = texts.AUTO_SYNC_CANCELLED[LANGUAGE]
+BATCH_SYNC_CANCELLED = texts.BATCH_SYNC_CANCELLED[LANGUAGE]
+NO_SYNC_PROCESS = texts.NO_SYNC_PROCESS[LANGUAGE]
+INVALID_SYNC_TOOL = texts.INVALID_SYNC_TOOL[LANGUAGE]
+FAILED_START_PROCESS = texts.FAILED_START_PROCESS[LANGUAGE]
+ERROR_OCCURRED = texts.ERROR_OCCURRED[LANGUAGE]
+ERROR_DECODING_SUBTITLE = texts.ERROR_DECODING_SUBTITLE[LANGUAGE]
+ERROR_EXECUTING_COMMAND = texts.ERROR_EXECUTING_COMMAND[LANGUAGE]
+DROP_VALID_FILES = texts.DROP_VALID_FILES[LANGUAGE]
+PAIRS_ADDED = texts.PAIRS_ADDED[LANGUAGE]
+UNPAIRED_FILES_ADDED = texts.UNPAIRED_FILES_ADDED[LANGUAGE]
+UNPAIRED_FILES = texts.UNPAIRED_FILES[LANGUAGE]
+DUPLICATE_PAIRS_SKIPPED = texts.DUPLICATE_PAIRS_SKIPPED[LANGUAGE]
+PAIR_ALREADY_EXISTS = texts.PAIR_ALREADY_EXISTS[LANGUAGE]
+PAIR_ADDED = texts.PAIR_ADDED[LANGUAGE]
+SAME_FILE_ERROR = texts.SAME_FILE_ERROR[LANGUAGE]
+PAIR_ALREADY_EXISTS = texts.PAIR_ALREADY_EXISTS[LANGUAGE]
+SUBTITLE_ADDED = texts.SUBTITLE_ADDED[LANGUAGE]
+VIDEO_ADDED = texts.VIDEO_ADDED[LANGUAGE]
+FILE_CHANGED = texts.FILE_CHANGED[LANGUAGE]
+SELECT_ITEM_TO_CHANGE = texts.SELECT_ITEM_TO_CHANGE[LANGUAGE]
+SELECT_ITEM_TO_REMOVE = texts.SELECT_ITEM_TO_REMOVE[LANGUAGE]
+FILE_NOT_EXIST = texts.FILE_NOT_EXIST[LANGUAGE]
+MULTIPLE_ARGUMENTS = texts.MULTIPLE_ARGUMENTS[LANGUAGE]
+INVALID_FILE_FORMAT = texts.INVALID_FILE_FORMAT[LANGUAGE]
+INVALID_SYNC_TOOL_SELECTED = texts.INVALID_SYNC_TOOL_SELECTED[LANGUAGE]
+TEXT_SELECTED_FOLDER = texts.TEXT_SELECTED_FOLDER[LANGUAGE]
+TEXT_NO_FOLDER_SELECTED = texts.TEXT_NO_FOLDER_SELECTED[LANGUAGE]
+TEXT_DESTINATION_FOLDER_DOES_NOT_EXIST = texts.TEXT_DESTINATION_FOLDER_DOES_NOT_EXIST[
+    LANGUAGE
+]
+ADDED_PAIRS_MSG = texts.ADDED_PAIRS_MSG[LANGUAGE]
+SKIPPED_DUPLICATES_MSG = texts.SKIPPED_DUPLICATES_MSG[LANGUAGE]
+NO_VALID_SUBTITLE_PAIRS_TO_PROCESS = texts.NO_VALID_SUBTITLE_PAIRS_TO_PROCESS[LANGUAGE]
+NO_MATCHING_SUBTITLE_PAIRS_FOUND = texts.NO_MATCHING_SUBTITLE_PAIRS_FOUND[LANGUAGE]
+NO_SUBTITLE_PAIRS_TO_PROCESS = texts.NO_SUBTITLE_PAIRS_TO_PROCESS[LANGUAGE]
+NOT_FIND_COMPATIBLE_FILE_MANAGER = texts.NOT_FIND_COMPATIBLE_FILE_MANAGER[LANGUAGE]
+FILE_NOT_FOUND = texts.FILE_NOT_FOUND[LANGUAGE]
+ERROR_OPENING_DIRECTORY = texts.ERROR_OPENING_DIRECTORY[LANGUAGE]
+CONFIG_FILE_NOT_FOUND = texts.CONFIG_FILE_NOT_FOUND[LANGUAGE]
+# Log window messages
+VIDEO_FILE_NOT_FOUND = texts.VIDEO_FILE_NOT_FOUND[LANGUAGE]
+FFPROBE_FAILED = texts.FFPROBE_FAILED[LANGUAGE]
+SUBTITLE_EXTRACTION_FAILED = texts.SUBTITLE_EXTRACTION_FAILED[LANGUAGE]
+SYNCING_LOG_TEXT = texts.SYNCING_LOG_TEXT[LANGUAGE]
+INVALID_PARENT_ITEM = texts.INVALID_PARENT_ITEM[LANGUAGE]
+SKIP_NO_VIDEO_NO_SUBTITLE = texts.SKIP_NO_VIDEO_NO_SUBTITLE[LANGUAGE]
+SKIP_NO_SUBTITLE = texts.SKIP_NO_SUBTITLE[LANGUAGE]
+SKIP_NO_VIDEO = texts.SKIP_NO_VIDEO[LANGUAGE]
+SKIP_UNPAIRED_ITEM = texts.SKIP_UNPAIRED_ITEM[LANGUAGE]
+SKIPPING_ALREADY_SYNCED = texts.SKIPPING_ALREADY_SYNCED[LANGUAGE]
+CONVERTING_SUBTITLE = texts.CONVERTING_SUBTITLE[LANGUAGE]
+ERROR_CONVERTING_SUBTITLE = texts.ERROR_CONVERTING_SUBTITLE[LANGUAGE]
+SUBTITLE_CONVERTED = texts.SUBTITLE_CONVERTED[LANGUAGE]
+ERROR_UNSUPPORTED_CONVERSION = texts.ERROR_UNSUPPORTED_CONVERSION[LANGUAGE]
+FAILED_CONVERT_SUBTITLE = texts.FAILED_CONVERT_SUBTITLE[LANGUAGE]
+FAILED_CONVERT_VIDEO = texts.FAILED_CONVERT_VIDEO[LANGUAGE]
+SPLIT_PENALTY_ZERO = texts.SPLIT_PENALTY_ZERO[LANGUAGE]
+SPLIT_PENALTY_SET = texts.SPLIT_PENALTY_SET[LANGUAGE]
+USING_REFERENCE_SUBTITLE = texts.USING_REFERENCE_SUBTITLE[LANGUAGE]
+USING_VIDEO_FOR_SYNC = texts.USING_VIDEO_FOR_SYNC[LANGUAGE]
+ENABLED_NO_FIX_FRAMERATE = texts.ENABLED_NO_FIX_FRAMERATE[LANGUAGE]
+ENABLED_GSS = texts.ENABLED_GSS[LANGUAGE]
+ADDITIONAL_ARGS_ADDED = texts.ADDITIONAL_ARGS_ADDED[LANGUAGE]
+SYNCING_STARTED = texts.SYNCING_STARTED[LANGUAGE]
+SYNCING_ENDED = texts.SYNCING_ENDED[LANGUAGE]
+SYNC_SUCCESS = texts.SYNC_SUCCESS[LANGUAGE]
+SYNC_ERROR = texts.SYNC_ERROR[LANGUAGE]
+SYNC_ERROR_OCCURRED = texts.SYNC_ERROR_OCCURRED[LANGUAGE]
+BATCH_SYNC_COMPLETED = texts.BATCH_SYNC_COMPLETED[LANGUAGE]
+PREPARING_SYNC = texts.PREPARING_SYNC[LANGUAGE]
+CONVERTING_TTML = texts.CONVERTING_TTML[LANGUAGE]
+CONVERTING_VTT = texts.CONVERTING_VTT[LANGUAGE]
+CONVERTING_SBV = texts.CONVERTING_SBV[LANGUAGE]
+CONVERTING_SUB = texts.CONVERTING_SUB[LANGUAGE]
+CONVERTING_ASS = texts.CONVERTING_ASS[LANGUAGE]
+CONVERTING_STL = texts.CONVERTING_STL[LANGUAGE]
+ERROR_CONVERTING_SUBTITLE = texts.ERROR_CONVERTING_SUBTITLE[LANGUAGE]
+CONVERSION_NOT_SUPPORTED = texts.CONVERSION_NOT_SUPPORTED[LANGUAGE]
+SUBTITLE_CONVERTED = texts.SUBTITLE_CONVERTED[LANGUAGE]
+RETRY_ENCODING_MSG = texts.RETRY_ENCODING_MSG[LANGUAGE]
+ALASS_SPEED_OPTIMIZATION_LOG = texts.ALASS_SPEED_OPTIMIZATION_LOG[LANGUAGE]
+ALASS_DISABLE_FPS_GUESSING_LOG = texts.ALASS_DISABLE_FPS_GUESSING_LOG[LANGUAGE]
+CHANGING_ENCODING_MSG = texts.CHANGING_ENCODING_MSG[LANGUAGE]
+ENCODING_CHANGED_MSG = texts.ENCODING_CHANGED_MSG[LANGUAGE]
+SYNC_SUCCESS_COUNT = texts.SYNC_SUCCESS_COUNT[LANGUAGE]
+SYNC_FAILURE_COUNT = texts.SYNC_FAILURE_COUNT[LANGUAGE]
+SYNC_FAILURE_COUNT_BATCH = texts.SYNC_FAILURE_COUNT_BATCH[LANGUAGE]
+ERROR_CHANGING_ENCODING_MSG = texts.ERROR_CHANGING_ENCODING_MSG[LANGUAGE]
+BACKUP_CREATED = texts.BACKUP_CREATED[LANGUAGE]
+CHECKING_SUBTITLE_STREAMS = texts.CHECKING_SUBTITLE_STREAMS[LANGUAGE]
+FOUND_COMPATIBLE_SUBTITLES = texts.FOUND_COMPATIBLE_SUBTITLES[LANGUAGE]
+NO_COMPATIBLE_SUBTITLES = texts.NO_COMPATIBLE_SUBTITLES[LANGUAGE]
+SUCCESSFULLY_EXTRACTED = texts.SUCCESSFULLY_EXTRACTED[LANGUAGE]
+CHOOSING_BEST_SUBTITLE = texts.CHOOSING_BEST_SUBTITLE[LANGUAGE]
+CHOSEN_SUBTITLE = texts.CHOSEN_SUBTITLE[LANGUAGE]
+FAILED_TO_EXTRACT_SUBTITLES = texts.FAILED_TO_EXTRACT_SUBTITLES[LANGUAGE]
+USED_THE_LONGEST_SUBTITLE = texts.USED_THE_LONGEST_SUBTITLE[LANGUAGE]
+DELETING_EXTRACTED_SUBTITLE_FOLDER = texts.DELETING_EXTRACTED_SUBTITLE_FOLDER[LANGUAGE]
+DELETING_CONVERTED_SUBTITLE = texts.DELETING_CONVERTED_SUBTITLE[LANGUAGE]
+ADDED_FILES_TEXT = texts.ADDED_FILES_TEXT[LANGUAGE]
+SKIPPED_DUPLICATE_FILES_TEXT = texts.SKIPPED_DUPLICATE_FILES_TEXT[LANGUAGE]
+SKIPPED_OTHER_LIST_FILES_TEXT = texts.SKIPPED_OTHER_LIST_FILES_TEXT[LANGUAGE]
+SKIPPED_SEASON_EPISODE_DUPLICATES_TEXT = texts.SKIPPED_SEASON_EPISODE_DUPLICATES_TEXT[
+    LANGUAGE
+]
+SKIPPED_INVALID_FORMAT_FILES_TEXT = texts.SKIPPED_INVALID_FORMAT_FILES_TEXT[LANGUAGE]
+NO_FILES_SELECTED = texts.NO_FILES_SELECTED[LANGUAGE]
+NO_ITEM_SELECTED_TO_REMOVE = texts.NO_ITEM_SELECTED_TO_REMOVE[LANGUAGE]
+NO_FILES_SELECTED_TO_SHOW_PATH = texts.NO_FILES_SELECTED_TO_SHOW_PATH[LANGUAGE]
+REMOVED_ITEM = texts.REMOVED_ITEM[LANGUAGE]
+FILES_MUST_CONTAIN_PATTERNS = texts.FILES_MUST_CONTAIN_PATTERNS[LANGUAGE]
+NO_VALID_SUBTITLE_FILES = texts.NO_VALID_SUBTITLE_FILES[LANGUAGE]
